refactor(ic-certified-assets): flatten system_context/ into a single file
`system_context/` held just two files — `mod.rs` (SystemContext, 62 lines)
and `canister_env.rs` (CanisterEnv, 60 lines) — that always travelled
together: SystemContext owns an Option<CanisterEnv> and is responsible for
lazily loading it. Splitting them across a submodule added navigation cost
without isolating anything. Merge into a single top-level
`system_context.rs`.

The only external reference to the inner path was a single tests.rs use of
`crate::system_context::canister_env::CanisterEnv`; updated alongside.
